# docs/swagger.yaml
components:
  schemas:
...
    BaseResponse:
      properties:
        msg:
          default: ''
          title: Msg
          type: string
        status:
          default: success
          title: Status
          type: string
      title: BaseResponse
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    UserUpdate:
      properties:
        email:
          anyOf:
          - format: email
            type: string
          - type: 'null'
          title: Email
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        is_enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Enabled
        is_user_agreement_accepted:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is User Agreement Accepted
        is_verified_email:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Verified Email
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        position:
          anyOf:
          - type: string
          - type: 'null'
          title: Position
        second_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Second Name
        username:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
        wiki_api_client_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Wiki Api Client Id
      title: UserUpdate
      type: object
...
  securitySchemes:
    WikiAccessToken:
      in: cookie
      name: wiki_access_token
      type: apiKey
    WikiApiKey:
      in: header
      name: X-Wiki-API-Key
      type: apiKey
    WikiBearer:
      scheme: bearer
      type: http
    WikiRootApiKey:
      description: 'Root key, allows you to perform actions as administrator, random
        user data is generated.

        **Limitation:** operations requiring an explicitly registered userare unavailable,
        example.: `Create: POST /api_key/`.'
      in: header
      name: X-Wiki-ROOT-API-Key
      type: apiKey
info:
  description: Welcome to Wiki's API documentation! Here you will able to discover
    all of the ways you can interact with the Wiki API.
  title: Wiki API
  version: 0.0.1
openapi: 3.1.0
paths:
...
  /api/v1/user/:
    delete:
      operationId: delete_user_api_v1_user__delete
      parameters:
      - in: query
        name: user_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: User Id
      - in: query
        name: username
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
      - in: query
        name: email
        required: false
        schema:
          anyOf:
          - format: email
            type: string
          - type: 'null'
          title: Email
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Delete user
      tags:
      - User
    put:
      operationId: update_user_api_v1_user__put
      parameters:
      - in: query
        name: user_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: User Id
      - in: query
        name: username
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
      - in: query
        name: email
        required: false
        schema:
          anyOf:
          - format: email
            type: string
          - type: 'null'
          title: Email
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserFullInfoResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Update user
      tags:
      - User
  /api/v1/user/all:
    get:
      operationId: get_users_api_v1_user_all_get
      parameters:
      - in: query
        name: username
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
      - in: query
        name: first_name
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
      - in: query
        name: last_name
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
      - in: query
        name: second_name
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Second Name
      - description: Page number
        in: query
        name: page
        required: false
        schema:
          default: 1
          description: Page number
          minimum: 1
          title: Page
          type: integer
      - description: Page size
        in: query
        name: size
        required: false
        schema:
          default: 50
          description: Page size
          maximum: 100
          minimum: 1
          title: Size
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_UserFullInfoResponse_'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Get all users
      tags:
      - User
  /api/v1/user/info:
    get:
      operationId: get_user_api_v1_user_info_get
      parameters:
      - in: query
        name: user_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: User Id
      - in: query
        name: username
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
      - in: query
        name: email
        required: false
        schema:
          anyOf:
          - format: email
            type: string
          - type: 'null'
          title: Email
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserFullInfoResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Get user by id or username or email
      tags:
      - User
...
  /api/v1/user_group/:
    delete:
      operationId: delete_user_group_api_v1_user_group__delete
      parameters:
      - in: query
        name: group_id
        required: true
        schema:
          format: uuid
          title: Group Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Delete group info by id
      tags:
      - UserGroup
    post:
      description: '## Create user group

        Available for users starting from the EDITOR responsibility level'
      operationId: create_user_group_api_v1_user_group__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroup'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupInfo'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Create user group
      tags:
      - UserGroup
  /api/v1/user_group/all:
    get:
      operationId: get_all_groups_info_api_v1_user_group_all_get
      parameters:
      - in: query
        name: name
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      - in: query
        name: description
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      - in: query
        name: is_members_can_add_to_group
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Members Can Add To Group
      - description: Page number
        in: query
        name: page
        required: false
        schema:
          default: 1
          description: Page number
          minimum: 1
          title: Page
          type: integer
      - description: Page size
        in: query
        name: size
        required: false
        schema:
          default: 50
          description: Page size
          maximum: 100
          minimum: 1
          title: Size
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_GroupInfo_'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Get all groups info
      tags:
      - UserGroup
  /api/v1/user_group/all/full-info:
    get:
      operationId: get_all_groups_full_info_api_v1_user_group_all_full_info_get
      parameters:
      - in: query
        name: name
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      - in: query
        name: description
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      - in: query
        name: is_members_can_add_to_group
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Members Can Add To Group
      - description: Page number
        in: query
        name: page
        required: false
        schema:
          default: 1
          description: Page number
          minimum: 1
          title: Page
          type: integer
      - description: Page size
        in: query
        name: size
        required: false
        schema:
          default: 50
          description: Page size
          maximum: 100
          minimum: 1
          title: Size
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_GroupInfoWithUsers_'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Get all groups info with members
      tags:
      - UserGroup
  /api/v1/user_group/full-info:
    get:
      description: '## Get user group info with members by id

        Get information about the group and all its members'
      operationId: get_user_group_full_info_api_v1_user_group_full_info_get
      parameters:
      - in: query
        name: group_id
        required: true
        schema:
          format: uuid
          title: Group Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupInfoWithUsers'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Get user group info with members by id
      tags:
      - UserGroup
  /api/v1/user_group/info:
    get:
      description: '## Get user group info by id

        Get base information without group members'
      operationId: get_user_group_info_api_v1_user_group_info_get
      parameters:
      - in: query
        name: group_id
        required: true
        schema:
          format: uuid
          title: Group Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupInfo'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Get user group info by id
      tags:
      - UserGroup
    put:
      operationId: update_user_group_info_api_v1_user_group_info_put
      parameters:
      - in: query
        name: group_id
        required: true
        schema:
          format: uuid
          title: Group Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupOptionalInfo'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupInfo'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Update group info by id
      tags:
      - UserGroup
  /api/v1/user_group/member:
    delete:
      operationId: remove_member_group_api_v1_user_group_member_delete
      parameters:
      - in: query
        name: group_id
        required: true
        schema:
          format: uuid
          title: Group Id
          type: string
      - in: query
        name: user_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: User Id
      - in: query
        name: username
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
      - in: query
        name: email
        required: false
        schema:
          anyOf:
          - format: email
            type: string
          - type: 'null'
          title: Email
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Remove member user from group
      tags:
      - UserGroup
    post:
      operationId: add_member_group_api_v1_user_group_member_post
      parameters:
      - in: query
        name: group_id
        required: true
        schema:
          format: uuid
          title: Group Id
          type: string
      - in: query
        name: user_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: User Id
      - in: query
        name: username
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
      - in: query
        name: email
        required: false
        schema:
          anyOf:
          - format: email
            type: string
          - type: 'null'
          title: Email
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Add member user for group
      tags:
      - UserGroup
  /api/v1/user_group/members:
    get:
      operationId: get_members_user_group_api_v1_user_group_members_get
      parameters:
      - in: query
        name: group_id
        required: true
        schema:
          format: uuid
          title: Group Id
          type: string
      - description: Page number
        in: query
        name: page
        required: false
        schema:
          default: 1
          description: Page number
          minimum: 1
          title: Page
          type: integer
      - description: Page size
        in: query
        name: size
        required: false
        schema:
          default: 50
          description: Page size
          maximum: 100
          minimum: 1
          title: Size
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_UserBaseInfoResponse_'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Get members group by id
      tags:
      - UserGroup
  /api/v1/user_group/user/full-info:
    get:
      operationId: get_groups_full_info_for_user_api_v1_user_group_user_full_info_get
      parameters:
      - in: query
        name: user_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: User Id
      - in: query
        name: username
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
      - in: query
        name: email
        required: false
        schema:
          anyOf:
          - format: email
            type: string
          - type: 'null'
          title: Email
      - in: query
        name: name
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      - in: query
        name: description
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      - in: query
        name: is_members_can_add_to_group
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Members Can Add To Group
      - description: Page number
        in: query
        name: page
        required: false
        schema:
          default: 1
          description: Page number
          minimum: 1
          title: Page
          type: integer
      - description: Page size
        in: query
        name: size
        required: false
        schema:
          default: 50
          description: Page size
          maximum: 100
          minimum: 1
          title: Size
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_GroupInfoWithUsers_'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Get groups info for user with members
      tags:
      - UserGroup
  /api/v1/user_group/user/info:
    get:
      operationId: get_groups_info_for_user_api_v1_user_group_user_info_get
      parameters:
      - in: query
        name: user_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: User Id
      - in: query
        name: username
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
      - in: query
        name: email
        required: false
        schema:
          anyOf:
          - format: email
            type: string
          - type: 'null'
          title: Email
      - in: query
        name: name
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      - in: query
        name: description
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      - in: query
        name: is_members_can_add_to_group
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Members Can Add To Group
      - description: Page number
        in: query
        name: page
        required: false
        schema:
          default: 1
          description: Page number
          minimum: 1
          title: Page
          type: integer
      - description: Page size
        in: query
        name: size
        required: false
        schema:
          default: 50
          description: Page size
          maximum: 100
          minimum: 1
          title: Size
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_GroupInfo_'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - WikiRootApiKey: []
      - WikiApiKey: []
      - WikiApiKey: []
      - WikiAccessToken: []
      - WikiBearer: []
      summary: Get groups info for user
      tags:
      - UserGroup
...
servers:
- description: Local server
  url: http://localhost:8000/
- description: Demo server
  url: https://api.wiki.example.com/
- description: Ngrok free server
  url: https://poorly-ideal-cobra.ngrok-free.app/